+++ exporter/importables.go
@@ -6,7 +6,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"log"
-	"path"
 	"regexp"
 	"strings"
 	"time"
@@ -21,7 +20,6 @@ import (
 	"github.com/databrickslabs/terraform-provider-databricks/workspace"
 
 	"github.com/databrickslabs/terraform-provider-databricks/storage"
-	"github.com/hashicorp/hcl/v2/hclsyntax"
 	"github.com/hashicorp/hcl/v2/hclwrite"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
@@ -134,7 +132,7 @@ var resourcesMap map[string]importable = map[string]importable{
 			name := "_" + s[len(s)-1] + "_" + fileNameMd5
 			return name
 		},
-		Body: func(ic *importContext, body *hclwrite.Body, r *resource) error {
+		Import: func(ic *importContext, r *resource) error {
 			dbfsAPI := storage.NewDbfsAPI(ic.Context, ic.Client)
 			content, err := dbfsAPI.Read(r.ID)
 			if err != nil {
@@ -146,16 +144,11 @@ var resourcesMap map[string]importable = map[string]importable{
 			if err != nil {
 				return err
 			}
-			// libraries installed with init scripts won't be exported.
-			b := body.AppendNewBlock("resource", []string{r.Resource, r.Name}).Body()
-			relativeFile := fmt.Sprintf("${path.module}/files/%s", fileName)
-			b.SetAttributeValue("path", cty.StringVal(strings.Replace(r.ID, "dbfs:", "", 1)))
-			b.SetAttributeRaw("source", hclwrite.Tokens{
-				&hclwrite.Token{Type: hclsyntax.TokenOQuote, Bytes: []byte{'"'}},
-				&hclwrite.Token{Type: hclsyntax.TokenQuotedLit, Bytes: []byte(relativeFile)},
-				&hclwrite.Token{Type: hclsyntax.TokenCQuote, Bytes: []byte{'"'}},
-			})
+			r.Data.Set("source", fileName)
 			return nil
+		},
+		Depends: []reference{
+			{Path: "source", File: true},
 		},
 	},
 	"databricks_instance_pool": {
@@ -276,6 +269,7 @@ var resourcesMap map[string]importable = map[string]importable{
 			{Path: "spark_python_task.python_file", Resource: "databricks_dbfs_file", Match: "dbfs_path"},
 			{Path: "spark_python_task.parameters", Resource: "databricks_dbfs_file", Match: "dbfs_path"},
 			{Path: "spark_jar_task.jar_uri", Resource: "databricks_dbfs_file", Match: "dbfs_path"},
+			{Path: "notebook_task.notebook_path", Resource: "databricks_notebook"},
 		},
 		Import: func(ic *importContext, r *resource) error {
 			var job jobs.JobSettings
@@ -322,6 +316,12 @@ var resourcesMap map[string]importable = map[string]importable{
 						r.Data.Set("library", libs)
 					}
 				}
+			}
+			if job.NotebookTask != nil {
+				ic.Emit(&resource{
+					Resource: "databricks_notebook",
+					ID:       job.NotebookTask.NotebookPath,
+				})
 			}
 			return ic.importLibraries(r.Data, s)
 		},
@@ -588,7 +588,8 @@ var resourcesMap map[string]importable = map[string]importable{
 			if scopes, err := ssAPI.List(); err == nil {
 				for i, scope := range scopes {
 					if !ic.MatchesName(scope.Name) {
-						log.Printf("[INFO] Secret scope %s doesn't match %s filter", scope.Name, ic.match)
+						log.Printf("[INFO] Secret scope %s doesn't match %s filter",
+							scope.Name, ic.match)
 						continue
 					}
 					ic.Emit(&resource{
@@ -596,8 +597,7 @@ var resourcesMap map[string]importable = map[string]importable{
 						ID:       scope.Name,
 						Name:     scope.Name,
 					})
-					log.Printf("[INFO] Imported %d of %d secret scopes",
-						i, len(scopes))
+					log.Printf("[INFO] Imported %d of %d secret scopes", i, len(scopes))
 				}
 			}
 			return nil
@@ -628,22 +628,15 @@ var resourcesMap map[string]importable = map[string]importable{
 	"databricks_secret": {
 		Service: "secrets",
 		Depends: []reference{
+			{Path: "string_value", Variable: true},
 			{Path: "scope", Resource: "databricks_secret_scope"},
 			{Path: "string_value", Resource: "vault_generic_secret", Match: "data"},
 			{Path: "string_value", Resource: "aws_kms_secrets", Match: "plaintext"},
 			{Path: "string_value", Resource: "azurerm_key_vault_secret", Match: "value"},
 			{Path: "string_value", Resource: "aws_secretsmanager_secret_version", Match: "secret_string"},
 		},
-		Body: func(ic *importContext, body *hclwrite.Body, r *resource) error {
-			b := body.AppendNewBlock("resource", []string{r.Resource, r.Name}).Body()
-			b.SetAttributeRaw("scope", ic.reference(ic.Importables[r.Resource],
-				[]string{"scope"}, r.Data.Get("scope").(string)))
-			// secret data is exposed only within notebooks
-			b.SetAttributeRaw("string_value", ic.variable(
-				r.Name, fmt.Sprintf("Secret %s from %s scope",
-					r.Data.Get("key"), r.Data.Get("scope"))))
-			b.SetAttributeValue("key", cty.StringVal(r.Data.Get("key").(string)))
-			return nil
+		Name: func(d *schema.ResourceData) string {
+			return fmt.Sprintf("%s_%s", d.Get("scope"), d.Get("key"))
 		},
 	},
 	"databricks_secret_acl": {
@@ -725,8 +718,7 @@ var resourcesMap map[string]importable = map[string]importable{
 			if name == "" {
 				return d.Id()
 			}
-			re := regexp.MustCompile(`[^0-9A-Za-z_]`)
-			return re.ReplaceAllString(name, "_")
+			return name
 		},
 		List: func(ic *importContext) error {
 			globalInitScripts, err := workspace.NewGlobalInitScriptsAPI(ic.Context, ic.Client).List()
@@ -742,7 +734,7 @@ var resourcesMap map[string]importable = map[string]importable{
 			}
 			return nil
 		},
-		Body: func(ic *importContext, body *hclwrite.Body, r *resource) error {
+		Import: func(ic *importContext, r *resource) error {
 			gis, err := workspace.NewGlobalInitScriptsAPI(ic.Context, ic.Client).Get(r.ID)
 			if err != nil {
 				return err
@@ -751,21 +743,15 @@ var resourcesMap map[string]importable = map[string]importable{
 			if err != nil {
 				return err
 			}
-			fileName, err := ic.createFile(path.Base(r.Name), content)
+			fileName, err := ic.createFile(fmt.Sprintf("%s.sh", r.Name), content)
 			log.Printf("Creating %s for %s", fileName, r)
 			if err != nil {
 				return err
 			}
-			relativeFile := fmt.Sprintf("${path.module}/files/%s", fileName)
-			b := body.AppendNewBlock("resource", []string{r.Resource, r.Name}).Body()
-			b.SetAttributeValue("name", cty.StringVal(gis.Name))
-			b.SetAttributeValue("enabled", cty.BoolVal(gis.Enabled))
-			b.SetAttributeRaw("source", hclwrite.Tokens{
-				&hclwrite.Token{Type: hclsyntax.TokenOQuote, Bytes: []byte{'"'}},
-				&hclwrite.Token{Type: hclsyntax.TokenQuotedLit, Bytes: []byte(relativeFile)},
-				&hclwrite.Token{Type: hclsyntax.TokenCQuote, Bytes: []byte{'"'}},
-			})
-			return nil
+			return r.Data.Set("source", fileName)
+		},
+		Depends: []reference{
+			{Path: "source", File: true},
 		},
 	},
 	"databricks_repo": {
@@ -774,11 +760,8 @@ var resourcesMap map[string]importable = map[string]importable{
 			name := d.Get("path").(string)
 			if name == "" {
 				return d.Id()
-			} else {
-				name = strings.TrimPrefix(name, "/")
 			}
-			re := regexp.MustCompile(`[^0-9A-Za-z_]`)
-			return re.ReplaceAllString(name, "_")
+			return strings.TrimPrefix(name, "/")
 		},
 		List: func(ic *importContext) error {
 			repoList, err := repos.NewReposAPI(ic.Context, ic.Client).ListAll()
@@ -806,47 +789,6 @@ var resourcesMap map[string]importable = map[string]importable{
 			}
 			return nil
 		},
-		Body: func(ic *importContext, body *hclwrite.Body, r *resource) error {
-			b := body.AppendNewBlock("resource", []string{r.Resource, r.Name}).Body()
-			b.SetAttributeValue("url", cty.StringVal(r.Data.Get("url").(string)))
-			b.SetAttributeValue("git_provider", cty.StringVal(r.Data.Get("git_provider").(string)))
-			t := r.Data.Get("branch").(string)
-			if t != "" {
-				b.SetAttributeValue("branch", cty.StringVal(t))
-			}
-			t = r.Data.Get("path").(string)
-			if t != "" {
-				b.SetAttributeValue("path", cty.StringVal(t))
-			}
-			return nil
-		},
-	},
-	"databricks_git_credential": {
-		Service: "repos",
-		Name: func(d *schema.ResourceData) string {
-			return d.Get("git_provider").(string) + "_" + d.Get("git_username").(string) + "_" + d.Id()
-		},
-		List: func(ic *importContext) error {
-			creds, err := repos.NewGitCredentialsAPI(ic.Context, ic.Client).List()
-			if err != nil {
-				return err
-			}
-			for offset, cred := range creds {
-				ic.Emit(&resource{
-					Resource: "databricks_git_credential",
-					ID:       fmt.Sprintf("%d", cred.ID),
-				})
-				log.Printf("[INFO] Scanned %d of %d Git credentials", offset+1, len(creds))
-			}
-			return nil
-		},
-		Body: func(ic *importContext, body *hclwrite.Body, r *resource) error {
-			b := body.AppendNewBlock("resource", []string{r.Resource, r.Name}).Body()
-			b.SetAttributeRaw("personal_access_token", ic.variable("git_token_"+r.Name, "Git token for "+r.Name))
-			b.SetAttributeValue("git_provider", cty.StringVal(r.Data.Get("git_provider").(string)))
-			b.SetAttributeValue("git_username", cty.StringVal(r.Data.Get("git_username").(string)))
-			return nil
-		},
 	},
 	"databricks_workspace_conf": {
 		Service: "workspace",
@@ -855,7 +797,11 @@ var resourcesMap map[string]importable = map[string]importable{
 		},
 		Import: func(ic *importContext, r *resource) error {
 			wsConfAPI := workspace.NewWorkspaceConfAPI(ic.Context, ic.Client)
-			keys := map[string]interface{}{"enableIpAccessLists": false, "maxTokenLifetimeDays": 0, "enableTokensConfig": false}
+			keys := map[string]interface{}{
+				"enableIpAccessLists":  false,
+				"maxTokenLifetimeDays": 0,
+				"enableTokensConfig":   false,
+			}
 			err := wsConfAPI.Read(&keys)
 			if err != nil {
 				return err
@@ -894,6 +840,65 @@ var resourcesMap map[string]importable = map[string]importable{
 				})
 			}
 			return nil
+		},
+	},
+	"databricks_notebook": {
+		Service: "notebooks",
+		Name: func(d *schema.ResourceData) string {
+			name := d.Get("path").(string)
+			if name == "" {
+				return d.Id()
+			}
+			return name
+		},
+		List: func(ic *importContext) error {
+			notebooksAPI := workspace.NewNotebooksAPI(ic.Context, ic.Client)
+			notebookList, err := notebooksAPI.List("/", true)
+			if err != nil {
+				return err
+			}
+			for offset, notebook := range notebookList {
+				if strings.HasPrefix(notebook.Path, "/Repos") {
+					continue
+				}
+				// TODO: emit permissions for notebook folders if non-default,
+				// as per-notebook permission entry would be a noise in the state
+				ic.Emit(&resource{
+					Resource: "databricks_notebook",
+					ID:       notebook.Path,
+				})
+				if offset%50 == 0 {
+					log.Printf("[INFO] Scanned %d of %d notebooks",
+						offset+1, len(notebookList))
+				}
+			}
+			return nil
+		},
+		Import: func(ic *importContext, r *resource) error {
+			notebooksAPI := workspace.NewNotebooksAPI(ic.Context, ic.Client)
+			contentB64, err := notebooksAPI.Export(r.ID, "SOURCE")
+			if err != nil {
+				return err
+			}
+			language := r.Data.Get("language").(string)
+			ext := map[string]string{
+				"SCALA":  ".scala",
+				"PYTHON": ".py",
+				"SQL":    ".sql",
+				"R":      ".r",
+			}
+			name := r.ID[1:] + ext[language] // todo: replace non-alphanum+/ with _
+			content, _ := base64.StdEncoding.DecodeString(contentB64)
+			fileName, err := ic.createFileIn("notebooks", name, []byte(content))
+			if err != nil {
+				return err
+			}
+			log.Printf("Creating %s for %s", fileName, r)
+			r.Data.Set("source", fileName)
+			return r.Data.Set("language", "")
+		},
+		Depends: []reference{
+			{Path: "source", File: true},
 		},
 	},
 }