+++ exporter/importables_test.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/databrickslabs/terraform-provider-databricks/clusters"
 	"github.com/databrickslabs/terraform-provider-databricks/common"
+	"github.com/databrickslabs/terraform-provider-databricks/internal"
 	"github.com/databrickslabs/terraform-provider-databricks/jobs"
 	"github.com/databrickslabs/terraform-provider-databricks/permissions"
 	"github.com/databrickslabs/terraform-provider-databricks/policies"
@@ -20,6 +21,7 @@ import (
 	"github.com/databrickslabs/terraform-provider-databricks/secrets"
 	"github.com/databrickslabs/terraform-provider-databricks/storage"
 	"github.com/databrickslabs/terraform-provider-databricks/workspace"
+	"github.com/hashicorp/hcl/v2/hclwrite"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -28,7 +30,9 @@ func importContextForTest() *importContext {
 	return &importContext{
 		Importables: resourcesMap,
 		Resources:   p.ResourcesMap,
+		Files:       map[string]*hclwrite.File{},
 		testEmits:   map[string]bool{},
+		nameFixes:   nameFixes,
 	}
 }
 
@@ -164,79 +168,6 @@ func TestSecretScope(t *testing.T) {
 	ic := importContextForTest()
 	name := ic.Importables["databricks_secret_scope"].Name(d)
 	assert.Equal(t, "abc", name)
-}
-
-func TestDbfsFileCornerCases_ReadFail(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/dbfs/read?length=1000000&path=a",
-			Status:   404,
-			Response: common.NotFound("nope"),
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTest()
-		ic.Client = client
-		ic.Context = ctx
-		err := resourcesMap["databricks_dbfs_file"].Body(ic, nil, &resource{
-			ID: "a",
-		})
-		assert.EqualError(t, err, "cannot read a: nope")
-	})
-}
-
-func TestDbfsFileCornerCases_WriteWrongDir(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/dbfs/read?length=1000000&path=a",
-			Response: storage.ReadResponse{
-				Data:      "YWJj",
-				BytesRead: 3,
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTest()
-		ic.Client = client
-		ic.Context = ctx
-		err := resourcesMap["databricks_dbfs_file"].Body(ic, nil, &resource{
-			ID:   "a",
-			Data: storage.ResourceDBFSFile().TestResourceData(),
-		})
-		assert.NotNil(t, err) // mustn't match direct OS error
-	})
-}
-
-func TestDbfsFileCornerCases_WriteFileExists(t *testing.T) {
-	qa.HTTPFixturesApply(t, []qa.HTTPFixture{
-		{
-			Method:   "GET",
-			Resource: "/api/2.0/dbfs/read?length=1000000&path=a",
-			Response: storage.ReadResponse{
-				Data:      "YWJj",
-				BytesRead: 3,
-			},
-		},
-	}, func(ctx context.Context, client *common.DatabricksClient) {
-		ic := importContextForTest()
-		ic.Client = client
-		ic.Context = ctx
-		ic.Directory = fmt.Sprintf("/tmp/tf-%s", qa.RandomName())
-		defer os.RemoveAll(ic.Directory)
-
-		dstFile := fmt.Sprintf("%s/files/_abc_900150983cd24fb0d6963f7d28e17f72", ic.Directory)
-		err := os.MkdirAll(dstFile, 0755)
-		assert.NoError(t, err)
-
-		d := storage.ResourceDBFSFile().TestResourceData()
-		d.SetId("abc")
-
-		err = resourcesMap["databricks_dbfs_file"].Body(ic, nil, &resource{
-			ID:   "a",
-			Data: d,
-		})
-		assert.Equal(t, err.Error(), fmt.Sprintf("open %s: is a directory", dstFile))
-	})
 }
 
 func TestInstancePoolNameFromID(t *testing.T) {
@@ -548,7 +479,7 @@ func TestGlobalInitScriptsErrors(t *testing.T) {
 		err := resourcesMap["databricks_global_init_script"].List(ic)
 		assert.EqualError(t, err, "nope")
 
-		err = resourcesMap["databricks_global_init_script"].Body(ic, nil, &resource{
+		err = resourcesMap["databricks_global_init_script"].Import(ic, &resource{
 			ID: "abc",
 		})
 		assert.EqualError(t, err, "nope")
@@ -577,12 +508,12 @@ func TestGlobalInitScriptsBodyErrors(t *testing.T) {
 		ic := importContextForTest()
 		ic.Client = client
 		ic.Context = ctx
-		err := resourcesMap["databricks_global_init_script"].Body(ic, nil, &resource{
+		err := resourcesMap["databricks_global_init_script"].Import(ic, &resource{
 			ID: "sad-emoji",
 		})
 		assert.EqualError(t, err, "illegal base64 data at input byte 0")
 
-		err = resourcesMap["databricks_global_init_script"].Body(ic, nil, &resource{
+		err = resourcesMap["databricks_global_init_script"].Import(ic, &resource{
 			ID: "second",
 		})
 		assert.NotNil(t, err) // no exact match because of OS diffs
@@ -609,5 +540,158 @@ func TestRepoListFails(t *testing.T) {
 		ic.Context = ctx
 		err := resourcesMap["databricks_repo"].List(ic)
 		assert.EqualError(t, err, "nope")
+	})
+}
+
+func testGenerate(t *testing.T, fixtures []qa.HTTPFixture, services string, cb func(*importContext)) {
+	qa.HTTPFixturesApply(t, fixtures, func(ctx context.Context, client *common.DatabricksClient) {
+		ic := importContextForTest()
+		ic.Directory = fmt.Sprintf("/tmp/tf-%s", qa.RandomName())
+		defer os.RemoveAll(ic.Directory)
+		ic.Client = client
+		ic.Context = ctx
+		ic.testEmits = nil
+		ic.importing = map[string]bool{}
+		ic.variables = map[string]string{}
+		ic.services = services
+		cb(ic)
+	})
+}
+
+func TestNotebookGeneration(t *testing.T) {
+	testGenerate(t, []qa.HTTPFixture{
+		{
+			Method:   "GET",
+			Resource: "/api/2.0/workspace/list?path=%2F",
+			Response: workspace.ObjectList{
+				Objects: []workspace.ObjectStatus{
+					{
+						Path:       "/Repos/Foo/Bar",
+						ObjectType: "NOTEBOOK",
+					},
+					{
+						Path:       "/First/Second",
+						ObjectType: "NOTEBOOK",
+					},
+				},
+			},
+		},
+		{
+			Method:   "GET",
+			Resource: "/api/2.0/workspace/get-status?path=%2FFirst%2FSecond",
+			Response: workspace.ObjectStatus{
+				ObjectID:   123,
+				ObjectType: "NOTEBOOK",
+				Path:       "/First/Second",
+				Language:   "PYTHON",
+			},
+		},
+		{
+			Method:   "GET",
+			Resource: "/api/2.0/workspace/export?format=SOURCE&path=%2FFirst%2FSecond",
+			Response: workspace.ExportPath{
+				Content: "YWJj",
+			},
+		},
+	}, "notebooks", func(ic *importContext) {
+		err := resourcesMap["databricks_notebook"].List(ic)
+		assert.NoError(t, err)
+
+		ic.generateHclForResources(nil)
+		assert.Equal(t, internal.TrimLeadingWhitespace(`
+		resource "databricks_notebook" "firstsecond" {
+		  source = "${path.module}/notebooks/First/Second.py"
+		  path   = "/First/Second"
+		}`), string(ic.Files["notebooks"].Bytes()))
+	})
+}
+
+func TestGlobalInitScriptGen(t *testing.T) {
+	testGenerate(t, []qa.HTTPFixture{
+		{
+			Method:       "GET",
+			ReuseRequest: true,
+			Resource:     "/api/2.0/global-init-scripts/a",
+			Response: workspace.GlobalInitScriptInfo{
+				Name:          "New: Importing ^ Things",
+				Enabled:       true,
+				ContentBase64: "YWJj",
+			},
+		},
+	}, "workspace", func(ic *importContext) {
+		ic.Emit(&resource{
+			Resource: "databricks_global_init_script",
+			ID:       "a",
+		})
+
+		ic.generateHclForResources(nil)
+		assert.Equal(t, internal.TrimLeadingWhitespace(`
+		resource "databricks_global_init_script" "new_importing_things" {
+		  source  = "${path.module}/files/new_importing_things.sh"
+		  name    = "New: Importing ^ Things"
+		  enabled = true
+		}`), string(ic.Files["workspace"].Bytes()))
+	})
+}
+
+func TestSecretGen(t *testing.T) {
+	testGenerate(t, []qa.HTTPFixture{
+		{
+			Method:   "GET",
+			Resource: "/api/2.0/secrets/list?scope=a",
+
+			Response: secrets.SecretsList{
+				Secrets: []secrets.SecretMetadata{
+					{
+						Key: "b",
+					},
+				},
+			},
+		},
+	}, "secrets", func(ic *importContext) {
+		ic.Emit(&resource{
+			Resource: "databricks_secret",
+			ID:       "a|||b",
+		})
+
+		ic.generateHclForResources(nil)
+		assert.Equal(t, internal.TrimLeadingWhitespace(`
+		resource "databricks_secret" "a_b" {
+		  string_value = var.string_value_a_b
+		  scope        = "a"
+		  key          = "b"
+		}`), string(ic.Files["secrets"].Bytes()))
+	})
+}
+
+func TestDbfsFileGen(t *testing.T) {
+	testGenerate(t, []qa.HTTPFixture{
+		{
+			Method:   "GET",
+			Resource: "/api/2.0/dbfs/get-status?path=a",
+			Response: storage.FileInfo{
+				Path: "a",
+			},
+		},
+		{
+			Method:   "GET",
+			Resource: "/api/2.0/dbfs/read?length=1000000&path=a",
+			Response: storage.ReadResponse{
+				Data:      "YWJj",
+				BytesRead: 3,
+			},
+		},
+	}, "storage", func(ic *importContext) {
+		ic.Emit(&resource{
+			Resource: "databricks_dbfs_file",
+			ID:       "a",
+		})
+
+		ic.generateHclForResources(nil)
+		assert.Equal(t, internal.TrimLeadingWhitespace(`
+		resource "databricks_dbfs_file" "_a_0cc175b9c0f1b6a831c399e269772661" {
+		  source = "${path.module}/files/_a_0cc175b9c0f1b6a831c399e269772661"
+		  path   = "a"
+		}`), string(ic.Files["storage"].Bytes()))
 	})
 }