+++ workspace/resource_notebook.go
@@ -110,6 +110,7 @@ func (a NotebooksAPI) Export(path string, format string) (string, error) {
 		Format: format,
 		Path:   path,
 	}, &notebookContent)
+	// TODO: return decoded []byte
 	return notebookContent.Content, err
 }
 
@@ -158,12 +159,12 @@ func (a NotebooksAPI) recursiveAddPaths(path string, pathList *[]ObjectStatus) e
 	return err
 }
 
-type objectList struct {
+type ObjectList struct {
 	Objects []ObjectStatus `json:"objects,omitempty"`
 }
 
 func (a NotebooksAPI) list(path string) ([]ObjectStatus, error) {
-	var notebookList objectList
+	var notebookList ObjectList
 	err := a.client.Get(a.context, "/workspace/list", map[string]string{
 		"path": path,
 	}, &notebookList)